% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoFreqTable.R
\name{PoDDyHePoFreqTable}
\alias{PoDDyHePoFreqTable}
\title{Frequency table Function Developed for PoDDy-Hepo Project}
\usage{
PoDDyHePoFreqTable(data, colName)
}
\arguments{
\item{data}{A data frame with items that need to be projected.}

\item{colName}{A column that you want to have a frequency table.}
}
\value{
a table
}
\description{
PoDDyHePoFreqTable returns a table arranged by Gender and Year.
}
\details{
Description of PoDDyHePoFreqTable()

The table includes Total, Missing and %.
}
