% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoModelSelection.R
\name{PoDDyHePoModelSelection}
\alias{PoDDyHePoModelSelection}
\title{Model Selection Function Developed for PoDDy-HePo Project (BIC based)}
\usage{
PoDDyHePoModelSelection(
  imp,
  DV,
  NsVar = NULL,
  df = NULL,
  knots = NULL,
  b.knots = NULL,
  f.var = NULL
)
}
\arguments{
\item{imp}{An object of class \code{mids} (produced by \code{mice()}.}

\item{DV}{Abbreviation of Dependent Variable.}

\item{NsVar}{A specific variable to which would apply nature splines. The default is \code{NULL}.}

\item{df}{Degrees of freedom. The default is \code{df=NULL}.}

\item{knots}{Breakpoints that define the spline.}

\item{b.knots}{Boundary points at which to impose the natural boundary conditions and anchor the B-spline basis (default the range of the data).}

\item{f.var}{Fixed variables. Variables that would like to fix it the model (will not be removed by stepwise process).}
}
\value{
Returns a list with 4 elements: Models, Mean BIC, Selected Models and Suggested Model.
}
\description{
PoDDyHePoModelSelection returns a list of 4 elements: 
             Models (Recoreds all selected models of each imputed data set), 
             Mean BIC (Records Mean BIC of each round of imputation),
             Selected Models (Records a selected model/selected models over imputation of cases)
             Suggested Model (Records a suggested model with smaller BIC value)
}
\details{
Description of PoDDyHePoModelSelection()

When NsVar = NULL, there is only one BIC value in Mean BIC and one model in Selected Models, the model in Selected Models is the same as in Suggested Model.
         df or knots should be deined.
         When both knots and b.knots are specified, df will be ignored.
}
