% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoPlot.R
\name{PoDDyHePoPlot}
\alias{PoDDyHePoPlot}
\title{Plot Function Developed for PoDDy-Hepo Project}
\usage{
PoDDyHePoPlot(data, year, title, sepvarlbl = NULL)
}
\arguments{
\item{data}{A data frame from PoDDyHePoPool.}

\item{year}{The lastest year that existing data.}

\item{title}{The tile of the plot. It changes due to we will have different reponse variable.}

\item{sepvarlbl}{Lable of separated variable. The default is \code{NULL}. Accoring to dummy_cols, the separated variables will be in the name for example obe_0, obe_1, obe_2,
if we taking obe for example. To label, sepvarlbl = c("Normal Weight", "Over Weight", "Obesity"). As the levels are sorted, so please make sure that the labels are in right order.}
}
\value{
Returns a graph
}
\description{
PoDDyHePoPlot returns a plot.
}
\details{
Description of PoDDyHePoPlot()

Enables plotting even there is a variable has been separated proportionally in the previous step.
}
