% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoPool.R
\name{PoDDyHePoPool}
\alias{PoDDyHePoPool}
\title{Pool Function Developed for Pooling Means and Standard Errors from Imputed Dataset.}
\usage{
PoDDyHePoPool(imp, colName)
}
\arguments{
\item{imp}{Complete data (mids object) from mice().}

\item{colName}{A variable that you want the estimated prevalence, Confidence Interval.}
}
\value{
Returns a data frame ready for plotting.
}
\description{
PoDDyHePoPool returns a data frame which is ready for plotting.
}
\details{
Description of PoDDyHePoPool()

It is able to detect separated variable and returns a data frame with Year, Sex, estimated prevalence (EST), CI_LOWER and CI_UPPER.
}
