% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoFreqTable.R
\name{PoDDyHePoFreqTable}
\alias{PoDDyHePoFreqTable}
\title{Frequency table Function Developed for PoDDy-Hepo Project}
\usage{
PoDDyHePoFreqTable(data, colName)
}
\arguments{
\item{data}{A data frame. It could be the well-formatted data, or the data frame whose prevalences are to be estimated.}

\item{colName}{A column that you want to create a frequency table for.}
}
\value{
a table
}
\description{
It returns a table arranged by Year and Sex.
}
\details{
Description of PoDDyHePoFreqTable()

The table includes total, respondents (N) and its proportion (%).
}
