% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoPlot.R
\name{PoDDyHePoPlot}
\alias{PoDDyHePoPlot}
\title{Plot Function Developed for PoDDy-Hepo Project}
\usage{
PoDDyHePoPlot(data, year, title, y_min, y_max, sepvarlbl = NULL)
}
\arguments{
\item{data}{A data frame from PoDDyHePoPool.}

\item{year}{The maximum year in the obseved data.}

\item{title}{The tile of the plot. It changes due to we will have different reponse variable.}

\item{y_min}{minimum value of y-axis(in percentage). For example, if 10 percent is the minimum value of y-axis, then type let y_min = 10.}

\item{y_max}{maximum value of y-axis(in percentage). Let y_max = 75, if 75 percent is the maximun value of y-axis.}

\item{sepvarlbl}{Lable of the variable whose porprotion of each category has been calculated. The default is \code{NULL}. Accoring to dummy_cols, the separated variables will be in the name for example obe_0, obe_1, obe_2, if we taking obe for example. To label, sepvarlbl = c("Normal Weight", "Over Weight", "Obesity"). As the levels are sorted, so please make sure that the labels are in right order.}
}
\value{
Returns a figure
}
\description{
It plots the prevalences against survey year.
}
\details{
Description of PoDDyHePoPlot()

Enables plotting even there is a variable whose porprotion of each category has been calculated.
}
