% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoPool.R
\name{PoDDyHePoPool}
\alias{PoDDyHePoPool}
\title{Pool Function Developed for PoDDy-HePo project.}
\usage{
PoDDyHePoPool(imp, colName, grpVar = NULL)
}
\arguments{
\item{imp}{Complete data (mids object) from PoDDyHePoProjeciton().}

\item{colName}{A variable whose prevalences are to be estimated.}

\item{grpVar}{Grouping variable. The default is NULL. In our case, "sex" used as grouping variable, that is grpVar = "sex".}
}
\value{
Returns a data frame ready for plotting.
}
\description{
It calculates prevalences for variables of interest and combines the estimates from imputed data set.
}
\details{
Description of PoDDyHePoPool()

It is able to detect separated variable and returns a data frame with Year, Sex, estimated prevalence (EST), CI_LOWER and CI_UPPER.
}
