% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoPopulationDF.R
\name{PoDDyHePoPopulationDF}
\alias{PoDDyHePoPopulationDF}
\title{Population Forecasts Function Developed for PoDDy-HePo Project}
\usage{
PoDDyHePoPopulationDF(data, file, size, y2pred)
}
\arguments{
\item{data}{Well-formatted data.}

\item{file}{Population forcasts from statistic Finland.}

\item{size}{Size of sample to be projected}

\item{y2pred}{years that need to be projected.}
}
\value{
a data frame
}
\description{
PoDDyHePoPopulationDF returns a data frame with data needs to be projected.
}
\details{
Description of PoDDyHePoPopulationDF()
}
