% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoDDyHePoProjection.R
\name{PoDDyHePoProjection}
\alias{PoDDyHePoProjection}
\title{Projection Function Developed for PoDDy-HePo Project}
\usage{
PoDDyHePoProjection(data, m = 5, maxit = 10, printFlag = F, ...)
}
\arguments{
\item{data}{A well-formatted data set ready for projection.}

\item{m}{Number of multiple imputations. The default is \code{m=5}.}

\item{maxit}{A scalar giving the number of iterations. By default, `maxit = 10`.}

\item{printFlag}{If `TRUE`, `mice` will print history on console. Use `print=FALSE` for silent computation.}

\item{...}{Formulas.}
}
\value{
Returns an S3 object of class \code{\link[=mids-class]{mids}}
       (multiply imputed data set)
}
\description{
PoDDyHePoProjection() is the main body for projection. It projects individual-level values of the variables using multiple imputation.
}
\details{
Description of PoDDyHePoProjection()

It allows user to set up basic settings like number of multiple imputations(\code{m}), number of iterations(\code{maxit}) and whether to show the history(\code{printFlag}). It returns a \code{\link[=mids-class]{mids}} object if and only if when every variable with missingess has given a formula.
}
